Update webpack2.yml
The job failed due to a missing module error where the WebAssembly file `src.wasm` and the directory `src` do not exist.

### Solution
Ensure that the `src.wasm` file and `src` directory exist and are correctly referenced in your Webpack configuration.

### Suggested Code Change
1. Verify that the `src.wasm` file is present in the `src` directory.
2. Update your Webpack configuration to correctly point to the `src.wasm` file.

Add the following to your `webpack.config.js`:

```javascript
module.exports = {
  // other configurations
  resolve: {
    extensions: ['.wasm', '.mjs', '.js', '.json']
  },
  module: {
    rules: [
      {
        test: /\.wasm$/,
        type: 'webassembly/async'
      }
    ]
  }
};
```

Ensure your project structure includes the `src` directory and the `src.wasm` file. If the file is generated, make sure the generation process is included in your build pipeline.